pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.28.0",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pyttsx3>=2.99",
    "reportlab>=4.4.3",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "speechrecognition>=3.14.3",
    "streamlit>=1.47.1",
    "weasyprint>=66.0",
]
